물류 06.19

+++ src/main/java/com/Erp/controller/logistics/AccountController.java
@@ -1,5 +1,6 @@
 package com.Erp.controller.logistics;
 
+import com.Erp.dto.UserDto;
 import com.Erp.dto.logistics.AccountAddDto;
 import com.Erp.dto.logistics.AccountFormDto;
 import com.Erp.entity.logistics.Account;
@@ -17,6 +18,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.security.Principal;
 import java.time.LocalDateTime;
@@ -70,19 +73,24 @@ public class AccountController {
     }
     //구매관리 거래처 페이지에서 거래처를 등록 또는 수정 요청을 하면 호출되는 메소드
     @PostMapping(value = "/accounts/addAccount")
-    public @ResponseBody ResponseEntity addBuyAccount(@RequestBody @Valid AccountAddDto data, BindingResult error, Principal principal){
+    public @ResponseBody ResponseEntity addBuyAccount(@RequestBody @Valid AccountAddDto data, BindingResult error, Principal principal, HttpServletRequest request){
+
+        boolean department = this.getSession(request);
+
+        if(!department){
+            return new ResponseEntity<String>("작성 권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
 
         StringBuilder erromessages = new StringBuilder();
-
         if (error.hasErrors()) {
             for (FieldError errorBean : error.getFieldErrors()) {
                 erromessages.append(errorBean.getDefaultMessage() + "\n");
             }
             return new ResponseEntity<String>(erromessages.toString(), HttpStatus.BAD_REQUEST);
         }
-        Member member  = memberService.getMemberName(principal.getName());
+        Member member = memberService.getMemberName(principal.getName());
 
-        if(data.getAcCode() == null || data.getAcCode().trim() == ""){//만약 코드에 유의미한 값이 들어 있다면 수정기능을 사용합니다.
+        if (data.getAcCode() == null || data.getAcCode().trim() == "") {//만약 코드에 유의미한 값이 들어 있다면 수정기능을 사용합니다.
             //여기는 등록 기능입니다.
             int count = 0;
 
@@ -90,16 +98,16 @@ public class AccountController {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");
             String strNowDate = regDateTime.format(formatter);
 
-            int year=Integer.parseInt(strNowDate.substring(0,4));
-            int month=Integer.parseInt(strNowDate.substring(4,6));
-            int day=Integer.parseInt(strNowDate.substring(6));
-            count = accountService.accountCount(year,month,day);
+            int year = Integer.parseInt(strNowDate.substring(0, 4));
+            int month = Integer.parseInt(strNowDate.substring(4, 6));
+            int day = Integer.parseInt(strNowDate.substring(6));
+            count = accountService.accountCount(year, month, day);
 
-            Account account = Account.setAccount(data,count,strNowDate,member.getName());
+            Account account = Account.setAccount(data, count, strNowDate, member.getName());
             accountService.saveAccount(account);
-        }else{
+        } else {
             //여기는 수정기능입니다.
-            accountService.updateBuy(data,member.getName());
+            accountService.updateBuy(data, member.getName());
         }
 
         return new ResponseEntity<>(HttpStatus.OK);
@@ -107,7 +115,13 @@ public class AccountController {
 
     //구매, 영업 거래처 페이지에 데이터 1건을 보내줍니다.
     @PostMapping(value = "/accounts/updateAccount")
-    public @ResponseBody ResponseEntity buyUpdateData(@RequestBody String code){
+    public @ResponseBody ResponseEntity buyUpdateData(@RequestBody String code, HttpServletRequest request){
+        
+        boolean department = this.getSession(request);
+
+        if(!department){
+            return new ResponseEntity<String>("수정 권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
         //현재 문제점 Ajax로 넘겨 받은 데이터안에 ""가 붙어있어서 문제 발생.
         code = code.substring(1,code.length()-1);
 
@@ -121,7 +135,12 @@ public class AccountController {
 
     //구매,판매 거래처를 삭제합니다.
     @PostMapping(value = "/accounts/deleteAccount")
-    public @ResponseBody ResponseEntity deleteBuyAccount(@RequestBody List<String> code) {
+    public @ResponseBody ResponseEntity deleteBuyAccount(@RequestBody List<String> code, HttpServletRequest request) {
+
+        boolean department = this.getSession(request);
+        if(!department){
+            return new ResponseEntity<String>("삭제 권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }
         accountService.deleteBuy(code);
         return new ResponseEntity<>( HttpStatus.OK);
     }
@@ -140,4 +159,18 @@ public class AccountController {
         return new ResponseEntity<String>(prDivCategory, HttpStatus.OK);
     }
 
+    public boolean getSession(HttpServletRequest request){
+
+        HttpSession session = request.getSession();
+        UserDto user = (UserDto) session.getAttribute("User");
+
+        boolean department = false;
+
+        if(user.getDepartment().equals("구매팀") || user.getDepartment().equals("영업팀")) {
+            department = true;
+            return department;
+        }else {
+            return department;
+        }
+    }
 }
